+++ NEWS.md
@@ -1,6 +1,15 @@
 Release History
 ===============
 
+Version 0.0.2 (2024-05-19)
+--------------------------
+
+- Add SerialNetwork trait to support multiple communication methods
+- Fix issue when receiving network timeout error
+- Add first steps to support raw serial communication
+- Add extra options to the CLI, allowing to publish in a loop
+
+
 Version 0.0.1 (2024-05-13)
 --------------------------